+++ src/nqm/irimager/irimager.cpp
@@ -6,6 +6,7 @@
 #include <pybind11/stl/filesystem.h>
 #include <pybind11/stl_bind.h>
 
+#include "./chrono.hpp"
 #include "./irimager_class.hpp"
 #include "./logger.hpp"
 #include "./logger_context_manager.hpp"
@@ -44,10 +45,15 @@ to control these cameras.)";
   // helps prevent deadlock when calling code that doesn't touch Python objs
   const auto no_gil = pybind11::call_guard<pybind11::gil_scoped_release>();
 
+  m.def("monotonic_to_system_clock", &nqm::irimager::clock_cast,
+        DOC(nqm, irimager, clock_cast), no_gil);
+
   pybind11::class_<IRImager>(m, "IRImager", DOC(IRImager))
       .def(pybind11::init<const std::filesystem::path &>(),
            DOC(IRImager, IRImager), no_gil)
       .def("get_frame", &IRImager::get_frame, DOC(IRImager, get_frame), no_gil)
+      .def("get_frame_monotonic", &IRImager::get_frame_monotonic,
+           DOC(IRImager, get_frame_monotonic), no_gil)
       .def("get_temp_range_decimal", &IRImager::get_temp_range_decimal,
            DOC(IRImager, get_temp_range_decimal), no_gil)
       .def("get_library_version", &IRImager::get_library_version,
@@ -65,6 +71,8 @@ to control these cameras.)";
            DOC(IRImager, IRImager), no_gil)
       .def("get_frame", &IRImagerMock::get_frame, DOC(IRImager, get_frame),
            no_gil)
+      .def("get_frame_monotonic", &IRImager::get_frame_monotonic,
+           DOC(IRImager, get_frame_monotonic), no_gil)
       .def("get_temp_range_decimal", &IRImagerMock::get_temp_range_decimal,
            DOC(IRImager, get_temp_range_decimal), no_gil)
       .def("start_streaming", &IRImagerMock::start_streaming,