+++ CHANGELOG.md
@@ -13,8 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   creates/deletes a
   [`nqm.logger.Logger`](https://nqminds.github.io/nqm-irimager/apidoc/nqm.irimager.html#nqm.irimager.Logger)
   object when used in a [`with:` statement][PEP 343] ([#81][]).
+- Add `nqm.irimager.IRImager.get_frame_monotonic` Python method, which can be
+  used to get the monotonic time of a frame directly from the
+  EvoCortex IRImagerDirect SDK ([#84][]).
+- Add `nqm.irimager.monotonic_to_system_clock` function to convert a monotonic
+  time to a system clock time ([#84][]).
 
 [#81]: https://github.com/nqminds/nqm-irimager/pull/81
+[#84]: https://github.com/nqminds/nqm-irimager/pull/84
 [PEP 343]: https://peps.python.org/pep-0343/
 
 ## [1.0.0] - 2023-10-30